devex-843 - refactoring the way we create users to not depend on try catch
