+++ shared/src/persistence/dynamo/users/createOrUpdateUser.js
@@ -78,7 +78,32 @@ exports.createUserRecords = async ({ applicationContext, user, userId }) => {
   };
 };
 
-exports.createUser = async ({
+const isUserAlreadyCreated = async ({
+  applicationContext,
+  email,
+  userPoolId,
+}) => {
+  try {
+    await applicationContext
+      .getCognito()
+      .adminGetUser({
+        UserPoolId: userPoolId,
+        Username: email,
+      })
+      .promise();
+    return true;
+  } catch (e) {
+    if (e.code === 'UserNotFoundException') {
+      return false;
+    } else {
+      throw e;
+    }
+  }
+};
+
+exports.isUserAlreadyCreated = isUserAlreadyCreated;
+
+exports.createOrUpdateUser = async ({
   applicationContext,
   disableCognitoUser = false,
   password,
@@ -90,7 +115,13 @@ exports.createUser = async ({
       ? process.env.USER_POOL_IRS_ID
       : process.env.USER_POOL_ID;
 
-  try {
+  const userExists = await isUserAlreadyCreated({
+    applicationContext,
+    email: user.email,
+    userPoolId,
+  });
+
+  if (!userExists) {
     const response = await applicationContext
       .getCognito()
       .adminCreateUser({
@@ -119,8 +150,7 @@ exports.createUser = async ({
       })
       .promise();
     userId = response.User.Username;
-  } catch (err) {
-    // the user already exists
+  } else {
     const response = await applicationContext
       .getCognito()
       .adminGetUser({
