+++ web-api/src/applicationContext.js
@@ -182,11 +182,14 @@ const {
   createNewPractitionerUser,
 } = require('../../shared/src/persistence/dynamo/users/createNewPractitionerUser');
 const {
+  createOrUpdatePractitionerUser,
+} = require('../../shared/src/persistence/dynamo/users/createOrUpdatePractitionerUser');
+const {
+  createOrUpdateUser,
+} = require('../../shared/src/persistence/dynamo/users/createOrUpdateUser');
+const {
   createPetitionerAccountInteractor,
 } = require('../../shared/src/business/useCases/users/createPetitionerAccountInteractor');
-const {
-  createPractitionerUser,
-} = require('../../shared/src/persistence/dynamo/users/createPractitionerUser');
 const {
   createPractitionerUserInteractor,
 } = require('../../shared/src/business/useCases/practitioners/createPractitionerUserInteractor');
@@ -202,9 +205,6 @@ const {
 const {
   createTrialSessionWorkingCopy,
 } = require('../../shared/src/persistence/dynamo/trialSessions/createTrialSessionWorkingCopy');
-const {
-  createUser,
-} = require('../../shared/src/persistence/dynamo/users/createUser');
 const {
   createUserForContact,
 } = require('../../shared/src/business/useCaseHelper/caseAssociation/createUserForContact');
@@ -1331,10 +1331,10 @@ const gatewayMethods = {
     createCaseDeadline,
     createCaseTrialSortMappingRecords,
     createMessage,
-    createPractitionerUser,
+    createOrUpdatePractitionerUser,
+    createOrUpdateUser,
     createTrialSession,
     createTrialSessionWorkingCopy,
-    createUser,
     deleteKeyCount,
     fetchPendingItems,
     getSesStatus,
